requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Admin==1.6.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
python-dotenv==1.0.0
Werkzeug==2.3.7
email-validator==2.1.0
pillow==10.0.1